refactor(planning): respond to PR comments
